# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/estimates": {
      "get": {
        "tags": [
          "estimates"
        ],
        "summary": "見積もり一覧取得",
        "operationId": "get_estimates_api_estimates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEstimatesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "estimates"
        ],
        "summary": "見積もり作成",
        "operationId": "create_estimate_api_estimates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
    "/api/estimates/{estimate_id}": {
      "get": {
        "tags": [
          "estimates"
        ],
        "summary": "見積もり詳細情報取得",
        "operationId": "get_estimate_api_estimates__estimate_id__get",
        "parameters": [
          {
            "name": "estimate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Estimate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "estimates"
        ],
        "summary": "見積もり更新",
        "operationId": "update_estimate_api_estimates__estimate_id__put",
        "parameters": [
          {
            "name": "estimate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Estimate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEstimateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "estimates"
        ],
        "summary": "見積もり削除",
        "operationId": "delete_estimate_api_estimates__estimate_id__delete",
        "parameters": [
          {
            "name": "estimate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Estimate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEstimateRequest": {
        "properties": {
          "project_name": {
            "type": "string",
            "title": "Project Name"
          }
        },
        "type": "object",
        "required": [
          "project_name"
        ],
        "title": "CreateEstimateRequest"
      },
      "CreateEstimateResponse": {
        "properties": {
          "estimate": {
            "$ref": "#/components/schemas/EstimateRead"
          }
        },
        "type": "object",
        "required": [
          "estimate"
        ],
        "title": "CreateEstimateResponse"
      },
...
      "EstimateItemCreate": {
        "properties": {
          "item_name": {
            "type": "string",
            "title": "Item Name",
            "description": "品名・摘要"
          },
          "price": {
            "type": "integer",
            "title": "Price",
            "description": "単価"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "description": "数量"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "description": "カテゴリID"
          }
        },
        "type": "object",
        "required": [
          "item_name",
          "price",
          "quantity",
          "category_id"
        ],
        "title": "EstimateItemCreate"
      },
      "EstimateItemRead": {
        "properties": {
          "item_name": {
            "type": "string",
            "title": "Item Name",
            "description": "品名・摘要"
          },
          "price": {
            "type": "integer",
            "title": "Price",
            "description": "単価"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "description": "数量"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "description": "カテゴリID"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "subtotal": {
            "type": "integer",
            "title": "Subtotal"
          },
          "category": {
            "$ref": "#/components/schemas/ItemCategoryRead"
          }
        },
        "type": "object",
        "required": [
          "item_name",
          "price",
          "quantity",
          "category_id",
          "id",
          "subtotal",
          "category"
        ],
        "title": "EstimateItemRead"
      },
      "EstimateListItem": {
        "properties": {
          "estimate_number": {
            "type": "string",
            "title": "Estimate Number",
            "description": "見積書番号"
          },
          "status": {
            "$ref": "#/components/schemas/EstimateStatus"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date",
            "description": "発行日"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date",
            "description": "有効期限"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "件名"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name",
            "description": "顧客名"
          },
          "in_charge_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Charge Name",
            "description": "担当者名"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "estimate_number",
          "status",
          "expiry_date",
          "project_name",
          "id"
        ],
        "title": "EstimateListItem"
      },
      "EstimateRead": {
        "properties": {
          "estimate_number": {
            "type": "string",
            "title": "Estimate Number",
            "description": "見積書番号"
          },
          "status": {
            "$ref": "#/components/schemas/EstimateStatus"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date",
            "description": "発行日"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date",
            "description": "有効期限"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "件名"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name",
            "description": "顧客名"
          },
          "in_charge_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Charge Name",
            "description": "担当者名"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EstimateItemRead"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "estimate_number",
          "status",
          "expiry_date",
          "project_name",
          "id"
        ],
        "title": "EstimateRead"
      },
      "EstimateStatus": {
        "type": "string",
        "enum": [
          "draft",
          "sent",
          "accepted",
          "rejected",
          "expired"
        ],
        "title": "EstimateStatus"
      },
...
      "GetEstimateResponse": {
        "properties": {
          "estimate": {
            "$ref": "#/components/schemas/EstimateRead"
          }
        },
        "type": "object",
        "required": [
          "estimate"
        ],
        "title": "GetEstimateResponse"
      },
      "GetEstimatesResponse": {
        "properties": {
          "estimates": {
            "items": {
              "$ref": "#/components/schemas/EstimateListItem"
            },
            "type": "array",
            "title": "Estimates"
          }
        },
        "type": "object",
        "required": [
          "estimates"
        ],
        "title": "GetEstimatesResponse"
      },
...
      "ItemCategoryRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "グループ名"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "is_deleted"
        ],
        "title": "ItemCategoryRead"
      },
...
      "UpdateEstimateRequest": {
        "properties": {
          "estimate_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate Number"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EstimateStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "in_charge_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Charge Name"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EstimateItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "title": "UpdateEstimateRequest"
      },
      "UpdateEstimateResponse": {
        "properties": {
          "estimate": {
            "$ref": "#/components/schemas/EstimateRead"
          }
        },
        "type": "object",
        "required": [
          "estimate"
        ],
        "title": "UpdateEstimateResponse"
      }
    }
  }
}